Randomari/requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
Django==3.2.7
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
